fix: logits -> fp32 through .float() for loss stability

+++ src/cdresearch/utils/train_test_val_lightning.py
@@ -37,14 +37,14 @@ class ChangeDetectionModel(pl.LightningModule):
     def training_step(self, batch, batch_idx):
         x, y = self.process_batch(batch)
         logits = self(x)
-        loss = self.loss_fn(logits, y)
+        loss = self.loss_fn(logits.float(), y)
         self.log("train_loss", loss, prog_bar=True, on_epoch=True, on_step=False)
         return loss
 
     def validation_step(self, batch, batch_idx):
         x, y = self.process_batch(batch)
         logits = self(x)
-        loss = self.loss_fn(logits, y)
+        loss = self.loss_fn(logits.float(), y)
         self.log("val_loss", loss, prog_bar=True, on_epoch=True, on_step=False)
         self.val_metrics.update(logits, y)
 
@@ -59,7 +59,7 @@ class ChangeDetectionModel(pl.LightningModule):
     def test_step(self, batch, batch_idx):
         x, y = self.process_batch(batch)
         logits = self(x)
-        loss = self.loss_fn(logits, y)
+        loss = self.loss_fn(logits.float(), y)
         self.log("test_loss", loss, prog_bar=True, on_epoch=True, on_step=True)
         self.test_metrics.update(logits, y)
